# api/docs/swagger.yaml
definitions:
...
  ek_user_service.City:
    properties:
      code:
        type: integer
      id:
        type: string
      name:
        type: string
      ru_name:
        type: string
      soato:
        type: integer
    type: object
...
  ek_user_service.CreateUpdateStaffSwag:
    properties:
      address:
        type: string
      city:
        $ref: '#/definitions/ek_user_service.City'
        type: object
      email:
        type: string
      extra_info:
        type: string
      first_name:
        type: string
      inn:
        type: integer
      last_name:
        type: string
      login:
        type: string
      middle_name:
        type: string
      organization_id:
        type: string
      passport_issue_date:
        type: string
      passport_issue_place:
        type: string
      passport_number:
        type: string
      password:
        type: string
      phone_number:
        type: string
      pinfl:
        type: string
      region:
        $ref: '#/definitions/ek_user_service.Region'
        type: object
      role_id:
        type: string
      status:
        type: boolean
    type: object
...
  ek_user_service.Region:
    properties:
      code:
        type: integer
      id:
        type: string
      name:
        type: string
      ru_name:
        type: string
      soato:
        type: integer
    type: object
...
  ek_variables.FailureResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  user_service.CreateResponse:
    properties:
      id:
        type: string
    type: object
...
  user_service.GetAllStaffsResponse:
    properties:
      count:
        type: integer
      staffs:
        items:
          $ref: '#/definitions/user_service.Staff'
        type: array
    type: object
  user_service.GetRole:
    properties:
      code:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      updated_at:
        type: string
    type: object
  user_service.Organization:
    properties:
      code:
        type: integer
      created_at:
        type: string
      description:
        type: string
      external_id:
        type: integer
      full_name:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      updated_at:
        type: string
    type: object
...
  user_service.Staff:
    properties:
      address:
        type: string
      city:
        $ref: '#/definitions/user_service.StaffCity'
        type: object
      created_at:
        type: string
      email:
        type: string
      external_id:
        type: integer
      first_name:
        type: string
      id:
        type: string
      inn:
        type: integer
      last_name:
        type: string
      login:
        type: string
      middle_name:
        type: string
      organization:
        $ref: '#/definitions/user_service.Organization'
        type: object
      passport_issue_date:
        type: string
      passport_issue_place:
        type: string
      passport_number:
        type: string
      password:
        type: string
      phone_number:
        type: string
      pinfl:
        type: string
      policy:
        type: integer
      region:
        $ref: '#/definitions/user_service.StaffRegion'
        type: object
      role:
        $ref: '#/definitions/user_service.GetRole'
        type: object
      soato:
        type: string
      status:
        type: boolean
      unique_name:
        type: string
      updated_at:
        type: string
      user_type:
        type: string
      verified:
        type: boolean
    type: object
  user_service.StaffCity:
    properties:
      code:
        type: integer
      id:
        type: string
      name:
        type: string
      ru_name:
        type: string
      soato:
        type: integer
    type: object
  user_service.StaffRegion:
    properties:
      code:
        type: integer
      id:
        type: string
      name:
        type: string
      ru_name:
        type: string
      soato:
        type: integer
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /v1/staff:
    get:
      consumes:
      - application/json
      description: API for getting all staff
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: organization_id
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: phone_number
        type: string
      - in: query
        name: role_id
        type: string
      - in: query
        name: search_string
        type: string
      - in: query
        name: soato
        type: string
      - in: query
        name: status
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.GetAllStaffsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Getting All Staffs
      tags:
      - staff
    post:
      consumes:
      - application/json
      description: API for creating staff
      parameters:
      - description: Staff
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/ek_user_service.CreateUpdateStaffSwag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user_service.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: staff
      tags:
      - staff
  /v1/staff-by-token:
    get:
      consumes:
      - application/json
      description: API for getting staff
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Staff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Staff by token
      tags:
      - staff
  /v1/staff/{staff_id}:
    get:
      consumes:
      - application/json
      description: API for getting staff
      parameters:
      - description: staff_id
        in: path
        name: staff_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Staff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Get Staff
      tags:
      - staff
    put:
      consumes:
      - application/json
      description: API for updating staff
      parameters:
      - description: staff_id
        in: path
        name: staff_id
        required: true
        type: string
      - description: staff
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/ek_user_service.CreateUpdateStaffSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Update Staff
      tags:
      - staff
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"